Manage Conflict
충돌 관리
