+++ [React] TodoCalendar/src/App.jsx
@@ -1,21 +1,23 @@
 import { useRef, useState } from "react"
 import TodoList from "./Components/Todo/TodoListPage"
 import Calendar from "./Components/Calendar/Calendar"
-import Calendar2 from "./Components/Calendar2"
+
 import LoginForm from "./Components/Login/LoginForm";
 import SignUpForm from "./Components/Login/SignUpForm";
+
 // import Time from "./Components/Time"
 
 
 function App() {
-  
 
   return (
     <div>
+
         {/* <Time /> */}
-        {/* <Calendar2 /> */}
         <LoginForm />
         <SignUpForm />
+        <Calendar />
+
         {/* <TodoList todos={ todos } onToggle={ onToggle } onDelete={ onDelete } onInsert={onInsert} /> */}
         {/* <button 
           onClick={() => {
