--- src/Components/Calendar/Calendar.jsx
@@ -1,12 +1,0 @@
-function Calendar({ todos }) {
-    const month = 4
-    const day = 30
-
-    return (
-        <div>
-            
-        </div>
-    );
-}
-
-export default Calendar;
